가족 리스트 API 기초작업

+++ domain-photo/src/main/kotlin/com/petdiary/domain/entity/Photo.kt
@@ -46,7 +46,7 @@ class Photo (
     @Column(name = "read_count", nullable = false)
     @ColumnDefault("0")
     @Comment("조회수")
-    var readCount: Integer = 0 as Integer,
+    var readCount: Int = 0,
 
     @Column(name = "is_default", nullable = false)
     @ColumnDefault("0")
